ahmad

+++ src/main/java/com/project1/wallet/WalletRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface WalletRepository extends JpaRepository<Wallet, Long> {
-    Optional<Wallet> findByUserId(Long id);
+    Optional<Wallet> findByUserId(int id);
     Optional<Wallet> findById(Long id);
 
 }
